+++ CMakeLists.txt
@@ -363,8 +363,6 @@ set(WEB_PLUGIN_FILES
         )
 
 set(API_PLUGIN_FILES
-        web/api/rrd2json.c
-        web/api/rrd2json.h
         web/api/web_api_v1.c
         web/api/web_api_v1.h
         web/api/badges/web_buffer_svg.c
@@ -395,6 +393,22 @@ set(API_PLUGIN_FILES
         web/api/queries/ses/ses.h
         web/api/queries/des/des.c
         web/api/queries/des/des.h
+        web/api/formatters/rrd2json.c
+        web/api/formatters/rrd2json.h
+        web/api/formatters/csv/csv.c
+        web/api/formatters/csv/csv.h
+        web/api/formatters/json/json.c
+        web/api/formatters/json/json.h
+        web/api/formatters/ssv/ssv.c
+        web/api/formatters/ssv/ssv.h
+        web/api/formatters/value/value.c
+        web/api/formatters/value/value.h
+        web/api/formatters/json_wrapper.c
+        web/api/formatters/json_wrapper.h
+        web/api/formatters/charts2json.c
+        web/api/formatters/charts2json.h
+        web/api/formatters/rrdset2json.c
+        web/api/formatters/rrdset2json.h
         )
 
 set(STREAMING_PLUGIN_FILES
