+++ web/api/formatters/Makefile.am
@@ -1,0 +1,15 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+AUTOMAKE_OPTIONS = subdir-objects
+MAINTAINERCLEANFILES = $(srcdir)/Makefile.in
+
+SUBDIRS = \
+    csv \
+    json \
+    ssv \
+    value \
+    $(NULL)
+
+dist_noinst_DATA = \
+	README.md \
+	$(NULL)
