--- web/api/rrd2json.c
@@ -1,695 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-#include "web_api_v1.h"
-
-void rrd_stats_api_v1_chart_with_data(RRDSET *st, BUFFER *wb, size_t *dimensions_count, size_t *memory_used)
-{
-    rrdset_rdlock(st);
-
-    buffer_sprintf(wb,
-        "\t\t{\n"
-        "\t\t\t\"id\": \"%s\",\n"
-        "\t\t\t\"name\": \"%s\",\n"
-        "\t\t\t\"type\": \"%s\",\n"
-        "\t\t\t\"family\": \"%s\",\n"
-        "\t\t\t\"context\": \"%s\",\n"
-        "\t\t\t\"title\": \"%s (%s)\",\n"
-        "\t\t\t\"priority\": %ld,\n"
-        "\t\t\t\"plugin\": \"%s\",\n"
-        "\t\t\t\"module\": \"%s\",\n"
-        "\t\t\t\"enabled\": %s,\n"
-        "\t\t\t\"units\": \"%s\",\n"
-        "\t\t\t\"data_url\": \"/api/v1/data?chart=%s\",\n"
-        "\t\t\t\"chart_type\": \"%s\",\n"
-        "\t\t\t\"duration\": %ld,\n"
-        "\t\t\t\"first_entry\": %ld,\n"
-        "\t\t\t\"last_entry\": %ld,\n"
-        "\t\t\t\"update_every\": %d,\n"
-        "\t\t\t\"dimensions\": {\n"
-        , st->id
-        , st->name
-        , st->type
-        , st->family
-        , st->context
-        , st->title, st->name
-        , st->priority
-        , st->plugin_name?st->plugin_name:""
-        , st->module_name?st->module_name:""
-        , rrdset_flag_check(st, RRDSET_FLAG_ENABLED)?"true":"false"
-        , st->units
-        , st->name
-        , rrdset_type_name(st->chart_type)
-        , st->entries * st->update_every
-        , rrdset_first_entry_t(st)
-        , rrdset_last_entry_t(st)
-        , st->update_every
-        );
-
-    unsigned long memory = st->memsize;
-
-    size_t dimensions = 0;
-    RRDDIM *rd;
-    rrddim_foreach_read(rd, st) {
-        if(rrddim_flag_check(rd, RRDDIM_FLAG_HIDDEN)) continue;
-
-        memory += rd->memsize;
-
-        buffer_sprintf(
-                wb
-                , "%s"
-                        "\t\t\t\t\"%s\": { \"name\": \"%s\" }"
-                , dimensions ? ",\n" : ""
-                , rd->id
-                , rd->name
-        );
-
-        dimensions++;
-    }
-
-    if(dimensions_count) *dimensions_count += dimensions;
-    if(memory_used) *memory_used += memory;
-
-    buffer_strcat(wb, "\n\t\t\t},\n\t\t\t\"green\": ");
-    buffer_rrd_value(wb, st->green);
-    buffer_strcat(wb, ",\n\t\t\t\"red\": ");
-    buffer_rrd_value(wb, st->red);
-
-    buffer_strcat(wb, ",\n\t\t\t\"alarms\": {\n");
-    size_t alarms = 0;
-    RRDCALC *rc;
-    for(rc = st->alarms; rc ; rc = rc->rrdset_next) {
-
-        buffer_sprintf(
-                wb
-                , "%s"
-                        "\t\t\t\t\"%s\": {\n"
-                        "\t\t\t\t\t\"id\": %u,\n"
-                        "\t\t\t\t\t\"status\": \"%s\",\n"
-                        "\t\t\t\t\t\"units\": \"%s\",\n"
-                        "\t\t\t\t\t\"update_every\": %d\n"
-                        "\t\t\t\t}"
-                , (alarms) ? ",\n" : ""
-                , rc->name
-                , rc->id
-                , rrdcalc_status2string(rc->status)
-                , rc->units
-                , rc->update_every
-        );
-
-        alarms++;
-    }
-
-    buffer_sprintf(wb,
-        "\n\t\t\t}\n\t\t}"
-        );
-
-    rrdset_unlock(st);
-}
-
-void rrd_stats_api_v1_chart(RRDSET *st, BUFFER *wb) {
-    rrd_stats_api_v1_chart_with_data(st, wb, NULL, NULL);
-}
-
-void rrd_stats_api_v1_charts(RRDHOST *host, BUFFER *wb) {
-    static char *custom_dashboard_info_js_filename = NULL;
-    size_t c, dimensions = 0, memory = 0, alarms = 0;
-    RRDSET *st;
-
-    time_t now = now_realtime_sec();
-
-    if(unlikely(!custom_dashboard_info_js_filename))
-        custom_dashboard_info_js_filename = config_get(CONFIG_SECTION_WEB, "custom dashboard_info.js", "");
-
-    buffer_sprintf(wb, "{\n"
-           "\t\"hostname\": \"%s\""
-        ",\n\t\"version\": \"%s\""
-        ",\n\t\"os\": \"%s\""
-        ",\n\t\"timezone\": \"%s\""
-        ",\n\t\"update_every\": %d"
-        ",\n\t\"history\": %ld"
-        ",\n\t\"custom_info\": \"%s\""
-        ",\n\t\"charts\": {"
-        , host->hostname
-        , host->program_version
-        , host->os
-        , host->timezone
-        , host->rrd_update_every
-        , host->rrd_history_entries
-        , custom_dashboard_info_js_filename
-        );
-
-    c = 0;
-    rrdhost_rdlock(host);
-    rrdset_foreach_read(st, host) {
-        if(rrdset_is_available_for_viewers(st)) {
-            if(c) buffer_strcat(wb, ",");
-            buffer_strcat(wb, "\n\t\t\"");
-            buffer_strcat(wb, st->id);
-            buffer_strcat(wb, "\": ");
-            rrd_stats_api_v1_chart_with_data(st, wb, &dimensions, &memory);
-
-            c++;
-            st->last_accessed_time = now;
-        }
-    }
-
-    RRDCALC *rc;
-    for(rc = host->alarms; rc ; rc = rc->next) {
-        if(rc->rrdset)
-            alarms++;
-    }
-    rrdhost_unlock(host);
-
-    buffer_sprintf(wb
-                   , "\n\t}"
-                    ",\n\t\"charts_count\": %zu"
-                    ",\n\t\"dimensions_count\": %zu"
-                    ",\n\t\"alarms_count\": %zu"
-                    ",\n\t\"rrd_memory_bytes\": %zu"
-                    ",\n\t\"hosts_count\": %zu"
-                    ",\n\t\"hosts\": ["
-                   , c
-                   , dimensions
-                   , alarms
-                   , memory
-                   , rrd_hosts_available
-    );
-
-    if(unlikely(rrd_hosts_available > 1)) {
-        rrd_rdlock();
-
-        size_t found = 0;
-        RRDHOST *h;
-        rrdhost_foreach_read(h) {
-            if(!rrdhost_should_be_removed(h, host, now)) {
-                buffer_sprintf(wb
-                               , "%s\n\t\t{"
-                                "\n\t\t\t\"hostname\": \"%s\""
-                                "\n\t\t}"
-                               , (found > 0) ? "," : ""
-                               , h->hostname
-                );
-
-                found++;
-            }
-        }
-
-        rrd_unlock();
-    }
-    else {
-        buffer_sprintf(wb
-                       , "\n\t\t{"
-                        "\n\t\t\t\"hostname\": \"%s\""
-                        "\n\t\t}"
-                       , host->hostname
-        );
-    }
-
-    buffer_sprintf(wb, "\n\t]\n}\n");
-}
-
-
-static void rrdr_buffer_print_format(BUFFER *wb, uint32_t format)  {
-    switch(format) {
-        case DATASOURCE_JSON:
-            buffer_strcat(wb, DATASOURCE_FORMAT_JSON);
-            break;
-
-        case DATASOURCE_DATATABLE_JSON:
-            buffer_strcat(wb, DATASOURCE_FORMAT_DATATABLE_JSON);
-            break;
-
-        case DATASOURCE_DATATABLE_JSONP:
-            buffer_strcat(wb, DATASOURCE_FORMAT_DATATABLE_JSONP);
-            break;
-
-        case DATASOURCE_JSONP:
-            buffer_strcat(wb, DATASOURCE_FORMAT_JSONP);
-            break;
-
-        case DATASOURCE_SSV:
-            buffer_strcat(wb, DATASOURCE_FORMAT_SSV);
-            break;
-
-        case DATASOURCE_CSV:
-            buffer_strcat(wb, DATASOURCE_FORMAT_CSV);
-            break;
-
-        case DATASOURCE_TSV:
-            buffer_strcat(wb, DATASOURCE_FORMAT_TSV);
-            break;
-
-        case DATASOURCE_HTML:
-            buffer_strcat(wb, DATASOURCE_FORMAT_HTML);
-            break;
-
-        case DATASOURCE_JS_ARRAY:
-            buffer_strcat(wb, DATASOURCE_FORMAT_JS_ARRAY);
-            break;
-
-        case DATASOURCE_SSV_COMMA:
-            buffer_strcat(wb, DATASOURCE_FORMAT_SSV_COMMA);
-            break;
-
-        default:
-            buffer_strcat(wb, "unknown");
-            break;
-    }
-}
-
-static void rrdr_json_wrapper_begin(RRDR *r, BUFFER *wb, uint32_t format, RRDR_OPTIONS options, int string_value) {
-    rrdset_check_rdlock(r->st);
-
-    long rows = rrdr_rows(r);
-    long c, i;
-    RRDDIM *rd;
-
-    //info("JSONWRAPPER(): %s: BEGIN", r->st->id);
-    char kq[2] = "",                    // key quote
-            sq[2] = "";                     // string quote
-
-    if( options & RRDR_OPTION_GOOGLE_JSON ) {
-        kq[0] = '\0';
-        sq[0] = '\'';
-    }
-    else {
-        kq[0] = '"';
-        sq[0] = '"';
-    }
-
-    buffer_sprintf(wb, "{\n"
-                       "   %sapi%s: 1,\n"
-                       "   %sid%s: %s%s%s,\n"
-                       "   %sname%s: %s%s%s,\n"
-                       "   %sview_update_every%s: %d,\n"
-                       "   %supdate_every%s: %d,\n"
-                       "   %sfirst_entry%s: %u,\n"
-                       "   %slast_entry%s: %u,\n"
-                       "   %sbefore%s: %u,\n"
-                       "   %safter%s: %u,\n"
-                       "   %sdimension_names%s: ["
-                   , kq, kq
-                   , kq, kq, sq, r->st->id, sq
-                   , kq, kq, sq, r->st->name, sq
-                   , kq, kq, r->update_every
-                   , kq, kq, r->st->update_every
-                   , kq, kq, (uint32_t)rrdset_first_entry_t(r->st)
-                   , kq, kq, (uint32_t)rrdset_last_entry_t(r->st)
-                   , kq, kq, (uint32_t)r->before
-                   , kq, kq, (uint32_t)r->after
-                   , kq, kq);
-
-    for(c = 0, i = 0, rd = r->st->dimensions; rd && c < r->d ;c++, rd = rd->next) {
-        if(unlikely(r->od[c] & RRDR_DIMENSION_HIDDEN)) continue;
-        if(unlikely((options & RRDR_OPTION_NONZERO) && !(r->od[c] & RRDR_DIMENSION_NONZERO))) continue;
-
-        if(i) buffer_strcat(wb, ", ");
-        buffer_strcat(wb, sq);
-        buffer_strcat(wb, rd->name);
-        buffer_strcat(wb, sq);
-        i++;
-    }
-    if(!i) {
-#ifdef NETDATA_INTERNAL_CHECKS
-        error("RRDR is empty for %s (RRDR has %d dimensions, options is 0x%08x)", r->st->id, r->d, options);
-#endif
-        rows = 0;
-        buffer_strcat(wb, sq);
-        buffer_strcat(wb, "no data");
-        buffer_strcat(wb, sq);
-    }
-
-    buffer_sprintf(wb, "],\n"
-                       "   %sdimension_ids%s: ["
-                   , kq, kq);
-
-    for(c = 0, i = 0, rd = r->st->dimensions; rd && c < r->d ;c++, rd = rd->next) {
-        if(unlikely(r->od[c] & RRDR_DIMENSION_HIDDEN)) continue;
-        if(unlikely((options & RRDR_OPTION_NONZERO) && !(r->od[c] & RRDR_DIMENSION_NONZERO))) continue;
-
-        if(i) buffer_strcat(wb, ", ");
-        buffer_strcat(wb, sq);
-        buffer_strcat(wb, rd->id);
-        buffer_strcat(wb, sq);
-        i++;
-    }
-    if(!i) {
-        rows = 0;
-        buffer_strcat(wb, sq);
-        buffer_strcat(wb, "no data");
-        buffer_strcat(wb, sq);
-    }
-
-    buffer_sprintf(wb, "],\n"
-                       "   %slatest_values%s: ["
-                   , kq, kq);
-
-    for(c = 0, i = 0, rd = r->st->dimensions; rd && c < r->d ;c++, rd = rd->next) {
-        if(unlikely(r->od[c] & RRDR_DIMENSION_HIDDEN)) continue;
-        if(unlikely((options & RRDR_OPTION_NONZERO) && !(r->od[c] & RRDR_DIMENSION_NONZERO))) continue;
-
-        if(i) buffer_strcat(wb, ", ");
-        i++;
-
-        storage_number n = rd->values[rrdset_last_slot(r->st)];
-
-        if(!does_storage_number_exist(n))
-            buffer_strcat(wb, "null");
-        else
-            buffer_rrd_value(wb, unpack_storage_number(n));
-    }
-    if(!i) {
-        rows = 0;
-        buffer_strcat(wb, "null");
-    }
-
-    buffer_sprintf(wb, "],\n"
-                       "   %sview_latest_values%s: ["
-                   , kq, kq);
-
-    i = 0;
-    if(rows) {
-        calculated_number total = 1;
-
-        if(unlikely(options & RRDR_OPTION_PERCENTAGE)) {
-            total = 0;
-            for(c = 0, rd = r->st->dimensions; rd && c < r->d ;c++, rd = rd->next) {
-                calculated_number *cn = &r->v[ (rrdr_rows(r) - 1) * r->d ];
-                calculated_number n = cn[c];
-
-                if(likely((options & RRDR_OPTION_ABSOLUTE) && n < 0))
-                    n = -n;
-
-                total += n;
-            }
-            // prevent a division by zero
-            if(total == 0) total = 1;
-        }
-
-        for(c = 0, i = 0, rd = r->st->dimensions; rd && c < r->d ;c++, rd = rd->next) {
-            if(unlikely(r->od[c] & RRDR_DIMENSION_HIDDEN)) continue;
-            if(unlikely((options & RRDR_OPTION_NONZERO) && !(r->od[c] & RRDR_DIMENSION_NONZERO))) continue;
-
-            if(i) buffer_strcat(wb, ", ");
-            i++;
-
-            calculated_number *cn = &r->v[ (rrdr_rows(r) - 1) * r->d ];
-            RRDR_VALUE_FLAGS *co = &r->o[ (rrdr_rows(r) - 1) * r->d ];
-            calculated_number n = cn[c];
-
-            if(co[c] & RRDR_VALUE_EMPTY) {
-                if(options & RRDR_OPTION_NULL2ZERO)
-                    buffer_strcat(wb, "0");
-                else
-                    buffer_strcat(wb, "null");
-            }
-            else {
-                if(unlikely((options & RRDR_OPTION_ABSOLUTE) && n < 0))
-                    n = -n;
-
-                if(unlikely(options & RRDR_OPTION_PERCENTAGE))
-                    n = n * 100 / total;
-
-                buffer_rrd_value(wb, n);
-            }
-        }
-    }
-    if(!i) {
-        rows = 0;
-        buffer_strcat(wb, "null");
-    }
-
-    buffer_sprintf(wb, "],\n"
-                       "   %sdimensions%s: %ld,\n"
-                       "   %spoints%s: %ld,\n"
-                       "   %sformat%s: %s"
-                   , kq, kq, i
-                   , kq, kq, rows
-                   , kq, kq, sq
-    );
-
-    rrdr_buffer_print_format(wb, format);
-
-    buffer_sprintf(wb, "%s,\n"
-                       "   %sresult%s: "
-                   , sq
-                   , kq, kq
-    );
-
-    if(string_value) buffer_strcat(wb, sq);
-    //info("JSONWRAPPER(): %s: END", r->st->id);
-}
-
-static void rrdr_json_wrapper_end(RRDR *r, BUFFER *wb, uint32_t format, uint32_t options, int string_value) {
-    (void)format;
-
-    char kq[2] = "",                    // key quote
-            sq[2] = "";                     // string quote
-
-    if( options & RRDR_OPTION_GOOGLE_JSON ) {
-        kq[0] = '\0';
-        sq[0] = '\'';
-    }
-    else {
-        kq[0] = '"';
-        sq[0] = '"';
-    }
-
-    if(string_value) buffer_strcat(wb, sq);
-
-    buffer_sprintf(wb, ",\n %smin%s: ", kq, kq);
-    buffer_rrd_value(wb, r->min);
-    buffer_sprintf(wb, ",\n %smax%s: ", kq, kq);
-    buffer_rrd_value(wb, r->max);
-    buffer_strcat(wb, "\n}\n");
-}
-
-int rrdset2value_api_v1(
-          RRDSET *st
-        , BUFFER *wb
-        , calculated_number *n
-        , const char *dimensions
-        , long points
-        , long long after
-        , long long before
-        , int group_method
-        , long group_time
-        , uint32_t options
-        , time_t *db_after
-        , time_t *db_before
-        , int *value_is_null
-) {
-    RRDR *r = rrd2rrdr(st, points, after, before, group_method, group_time, options, dimensions);
-    if(!r) {
-        if(value_is_null) *value_is_null = 1;
-        return 500;
-    }
-
-    if(rrdr_rows(r) == 0) {
-        rrdr_free(r);
-
-        if(db_after)  *db_after  = 0;
-        if(db_before) *db_before = 0;
-        if(value_is_null) *value_is_null = 1;
-
-        return 400;
-    }
-
-    if(wb) {
-        if (r->result_options & RRDR_RESULT_OPTION_RELATIVE)
-            buffer_no_cacheable(wb);
-        else if (r->result_options & RRDR_RESULT_OPTION_ABSOLUTE)
-            buffer_cacheable(wb);
-    }
-
-    if(db_after)  *db_after  = r->after;
-    if(db_before) *db_before = r->before;
-
-    long i = (!(options & RRDR_OPTION_REVERSED))?rrdr_rows(r) - 1:0;
-    *n = rrdr2value(r, i, options, value_is_null);
-
-    rrdr_free(r);
-    return 200;
-}
-
-int rrdset2anything_api_v1(
-          RRDSET *st
-        , BUFFER *wb
-        , BUFFER *dimensions
-        , uint32_t format
-        , long points
-        , long long after
-        , long long before
-        , int group_method
-        , long group_time
-        , uint32_t options
-        , time_t *latest_timestamp
-) {
-    st->last_accessed_time = now_realtime_sec();
-
-    RRDR *r = rrd2rrdr(st, points, after, before, group_method, group_time, options, dimensions?buffer_tostring(dimensions):NULL);
-    if(!r) {
-        buffer_strcat(wb, "Cannot generate output with these parameters on this chart.");
-        return 500;
-    }
-
-    if(r->result_options & RRDR_RESULT_OPTION_RELATIVE)
-        buffer_no_cacheable(wb);
-    else if(r->result_options & RRDR_RESULT_OPTION_ABSOLUTE)
-        buffer_cacheable(wb);
-
-    if(latest_timestamp && rrdr_rows(r) > 0)
-        *latest_timestamp = r->before;
-
-    switch(format) {
-    case DATASOURCE_SSV:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 1);
-            rrdr2ssv(r, wb, options, "", " ", "");
-            rrdr_json_wrapper_end(r, wb, format, options, 1);
-        }
-        else {
-            wb->contenttype = CT_TEXT_PLAIN;
-            rrdr2ssv(r, wb, options, "", " ", "");
-        }
-        break;
-
-    case DATASOURCE_SSV_COMMA:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 1);
-            rrdr2ssv(r, wb, options, "", ",", "");
-            rrdr_json_wrapper_end(r, wb, format, options, 1);
-        }
-        else {
-            wb->contenttype = CT_TEXT_PLAIN;
-            rrdr2ssv(r, wb, options, "", ",", "");
-        }
-        break;
-
-    case DATASOURCE_JS_ARRAY:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-            rrdr2ssv(r, wb, options, "[", ",", "]");
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        }
-        else {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr2ssv(r, wb, options, "[", ",", "]");
-        }
-        break;
-
-    case DATASOURCE_CSV:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 1);
-            rrdr2csv(r, wb, options, "", ",", "\\n", "");
-            rrdr_json_wrapper_end(r, wb, format, options, 1);
-        }
-        else {
-            wb->contenttype = CT_TEXT_PLAIN;
-            rrdr2csv(r, wb, options, "", ",", "\r\n", "");
-        }
-        break;
-
-    case DATASOURCE_CSV_JSON_ARRAY:
-        wb->contenttype = CT_APPLICATION_JSON;
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-            buffer_strcat(wb, "[\n");
-            rrdr2csv(r, wb, options + RRDR_OPTION_LABEL_QUOTES, "[", ",", "]", ",\n");
-            buffer_strcat(wb, "\n]");
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        }
-        else {
-            wb->contenttype = CT_TEXT_PLAIN;
-            buffer_strcat(wb, "[\n");
-            rrdr2csv(r, wb, options + RRDR_OPTION_LABEL_QUOTES, "[", ",", "]", ",\n");
-            buffer_strcat(wb, "\n]");
-        }
-        break;
-
-    case DATASOURCE_TSV:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 1);
-            rrdr2csv(r, wb, options, "", "\t", "\\n", "");
-            rrdr_json_wrapper_end(r, wb, format, options, 1);
-        }
-        else {
-            wb->contenttype = CT_TEXT_PLAIN;
-            rrdr2csv(r, wb, options, "", "\t", "\r\n", "");
-        }
-        break;
-
-    case DATASOURCE_HTML:
-        if(options & RRDR_OPTION_JSON_WRAP) {
-            wb->contenttype = CT_APPLICATION_JSON;
-            rrdr_json_wrapper_begin(r, wb, format, options, 1);
-            buffer_strcat(wb, "<html>\\n<center>\\n<table border=\\\"0\\\" cellpadding=\\\"5\\\" cellspacing=\\\"5\\\">\\n");
-            rrdr2csv(r, wb, options, "<tr><td>", "</td><td>", "</td></tr>\\n", "");
-            buffer_strcat(wb, "</table>\\n</center>\\n</html>\\n");
-            rrdr_json_wrapper_end(r, wb, format, options, 1);
-        }
-        else {
-            wb->contenttype = CT_TEXT_HTML;
-            buffer_strcat(wb, "<html>\n<center>\n<table border=\"0\" cellpadding=\"5\" cellspacing=\"5\">\n");
-            rrdr2csv(r, wb, options, "<tr><td>", "</td><td>", "</td></tr>\n", "");
-            buffer_strcat(wb, "</table>\n</center>\n</html>\n");
-        }
-        break;
-
-    case DATASOURCE_DATATABLE_JSONP:
-        wb->contenttype = CT_APPLICATION_X_JAVASCRIPT;
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-
-        rrdr2json(r, wb, options, 1);
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        break;
-
-    case DATASOURCE_DATATABLE_JSON:
-        wb->contenttype = CT_APPLICATION_JSON;
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-
-        rrdr2json(r, wb, options, 1);
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        break;
-
-    case DATASOURCE_JSONP:
-        wb->contenttype = CT_APPLICATION_X_JAVASCRIPT;
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-
-        rrdr2json(r, wb, options, 0);
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        break;
-
-    case DATASOURCE_JSON:
-    default:
-        wb->contenttype = CT_APPLICATION_JSON;
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_begin(r, wb, format, options, 0);
-
-        rrdr2json(r, wb, options, 0);
-
-        if(options & RRDR_OPTION_JSON_WRAP)
-            rrdr_json_wrapper_end(r, wb, format, options, 0);
-        break;
-    }
-
-    rrdr_free(r);
-    return 200;
-}
